requirements.txt
amqp==5.3.1
asgiref==3.11.1
billiard==4.2.4
celery==5.6.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==46.0.7
Django==6.0.4
django-allauth==65.15.1
django-cors-headers==4.9.0
djangorestframework==3.17.1
google-api-core==2.30.3
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-auth-oauthlib==1.3.1
googleapis-common-protos==1.74.0
httplib2==0.31.2
idna==3.11
kombu==5.6.2
oauthlib==3.3.1
packaging==26.0
prompt_toolkit==3.0.52
proto-plus==1.27.2
protobuf==7.34.1
psycopg2-binary==2.9.11
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
redis==7.4.0
requests==2.33.1
requests-oauthlib==2.0.0
six==1.17.0
sqlparse==0.5.5
tzdata==2026.1
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
